New look and feel

+++ src/ToDo.Core/MongoDB/MongoContext.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using MongoDB.Driver;
 using ToDo.Core.Models;
 using MongoDB.Bson;
+using ToDo.Core.Service;
 
 namespace ToDo.Core.MondoDB
 {
@@ -13,11 +14,13 @@ namespace ToDo.Core.MondoDB
         private MongoClient _client;
         private IMongoDatabase _db;
         private IMongoCollection<ToDoItem> _collection;
-        public MongoContext()
+        private string  _userName;
+        public MongoContext(IUserService userService)
         {
             _client = new MongoClient("mongodb://localhost:27017");
             _db = _client.GetDatabase("Jeevan");
             _collection= _db.GetCollection<ToDoItem>("TODO");
+            _userName = userService?.UserName;
         }
 
         internal void Add(ToDoItem item)
@@ -27,12 +30,12 @@ namespace ToDo.Core.MondoDB
 
         internal ToDoItem FindById(Guid id)
         {
-            return _collection.Find<ToDoItem>(c => c.Id == id).FirstOrDefault();
+            return _collection.Find<ToDoItem>(c => c.Id == id && c.CreatedBy.Equals(this._userName, StringComparison.OrdinalIgnoreCase)).FirstOrDefault();
         }
 
         internal void Remove(ToDoItem match)
         {
-            if(match!=null)
+            if(match!=null && match.CreatedBy.Equals(this._userName, StringComparison.OrdinalIgnoreCase))
             {
                 _collection.DeleteOne(c => c.Id == match.Id);
             }
@@ -41,12 +44,12 @@ namespace ToDo.Core.MondoDB
 
         internal IEnumerable<ToDoItem> GetAll()
         {
-            return _collection.Find<ToDoItem>(new BsonDocument()).ToList<ToDoItem>();
+            return _collection.Find<ToDoItem>(c=> c.CreatedBy.Equals(this._userName, StringComparison.OrdinalIgnoreCase)).ToList<ToDoItem>();
         }
 
         internal IEnumerable<ToDoItem> GetCompleted()
         {
-            return _collection.Find<ToDoItem>(c => c.IsComplete == true).ToList<ToDoItem>();
+            return _collection.Find<ToDoItem>(c => c.IsComplete == true && c.CreatedBy.Equals(this._userName, StringComparison.OrdinalIgnoreCase)).ToList<ToDoItem>();
         }
 
         internal void Update(ToDoItem item)
@@ -56,12 +59,12 @@ namespace ToDo.Core.MondoDB
                                                    .Set(nameof(item.IsComplete), item.IsComplete)
                                                    .Set(nameof(item.UpdatedBy), item.UpdatedBy)
                                                    .Set(nameof(item.UpdatedDate), item.UpdatedDate);
-            _collection.UpdateOne<ToDoItem>(c => c.Id == item.Id, update);
+            _collection.UpdateOne<ToDoItem>(c => c.Id == item.Id &&  c.CreatedBy.Equals(this._userName, StringComparison.OrdinalIgnoreCase), update);
         }
 
         internal IEnumerable<ToDoItem> GetActive()
         {
-            return _collection.Find<ToDoItem>(c=> c.IsComplete==false).ToList<ToDoItem>();
+            return _collection.Find<ToDoItem>(c=> c.IsComplete==false &&  c.CreatedBy.Equals(this._userName, StringComparison.OrdinalIgnoreCase)).ToList<ToDoItem>();
         }
     }
 }
